Refactor element fragment and adapter names
